More refactoring into gamestate folders
