+++ src/actions/mod.rs
@@ -1,12 +1,12 @@
 use bevy::prelude::*;
 
-use crate::GameState;
-
 pub mod resources;
 mod systems;
 
 use resources::*;
 use systems::*;
+
+use crate::states::GameState;
 
 pub struct ActionsPlugin;
 
@@ -16,6 +16,6 @@ impl Plugin for ActionsPlugin {
     fn build(&self, app: &mut App) {
         app.init_resource::<Actions>()
             .add_system(handle_escape)
-            .add_system(set_movement_actions.in_set(OnUpdate(GameState::Playing)));
+            .add_system(set_movement_actions.in_set(OnUpdate(GameState::Game)));
     }
 }
