+++ src/game/physics/systems.rs
@@ -1,9 +1,6 @@
 use bevy::prelude::*;
 use bevy_rapier2d::prelude::*;
 
-#[derive(Component)]
-pub struct Walls;
-
 pub fn display_collision_events(
     mut collision_events: EventReader<CollisionEvent>,
     mut contact_force_events: EventReader<ContactForceEvent>,
