+++ src/game/player/systems.rs
@@ -3,10 +3,10 @@ use super::events::*;
 use super::{PLAYER_SIZE, PLAYER_SPEED};
 use crate::actions::resources::Actions;
 use crate::asset_loader::resources::*;
-use crate::enemy::components::Enemy;
-use crate::events::GameOverEvent;
-use crate::resources::Score;
-use crate::star::components::Star;
+use crate::game::enemy::components::Enemy;
+use crate::game::events::GameOverEvent;
+use crate::game::resources::Score;
+use crate::game::star::components::Star;
 use bevy::prelude::*;
 use bevy::window::PrimaryWindow;
 use bevy_rapier2d::prelude::*;
