+++ src/game/star/components.rs
