+++ src/game/star/resources.rs
