+++ src/game/star/systems.rs
@@ -2,7 +2,7 @@ use super::components::*;
 use super::resources::*;
 use super::STAR_SIZE;
 use crate::asset_loader::resources::*;
-use crate::walls::WALLS_MARGIN;
+use crate::game::walls::WALLS_MARGIN;
 use bevy::prelude::*;
 use bevy::window::PrimaryWindow;
 use bevy_rapier2d::prelude::*;
