+++ src/game/walls/components.rs
